Commence tutorial
This adds a tutorial section to the documentation, which is a
relative rehash of other sections but aims to tell an improved
story on how to use the Operator. This is still a work in progress
and more effort will be put into it prior to the 4.5.0 GA, but
it is good enough to get started.