Proper Deletion Behavior for Logical Backup PVCs
Logical backup PVCs are now properly identified when rmdata is run.
This ensures any PVCs for logical backups are properly preserved or
deleted depending on the specific settings specified when deleting a
cluster.  For instance, if '--keep-data' only is specified, then the
logical backups will now be deleted.  Additionally, if '--keep-backups'
is specified, then the logical backup PVCs will now be preserved.  This
is specifically accomplished by now matching on a prefix for the name
of any logical backup PVCs (prefix 'backup-<clusterName>-pgdump'),
instead of matching on the full name of the PVC (which includes the
name of the database name used for the logical backup, e.g.
'backup-<clusterName>-pgdump-<database>-pvc', and therefore could
vary).

Issue: [ch9381]