Remove Invalid pgcluster Already Processed Checks
Within the pgcluster controller, the check for an existing Kubernetes
Deployment matching the name of the pgcluster, as well as the check to
determine whether or not a pgcluster custom resource is in a
'processed' status prior to enqueuing it from the pgcluster
controller's 'onAdd' function, have been removed, since both are no
longer valid methods for determining whether or not to process a
pgcluster.  For instance, a Deployment matching the pgcluster name
(i.e. the original primary for the cluster) might no longer exist if it
was since scaled down and removed, while the 'processed' status does
not properly indicate whether or not the various Kubernetes resources
(e.g. Deployments) for the initial primary utilized to bootstrap the
cluster have been created, nor does it properly indicate whether or not
the PG cluster has been fully initialized.

Issue: [ch9574]