pgo show backup Command Update
Currently, the 'pgo show backup' functionality requires the
Primary pod to be in a Ready state. This is not technically required
because the relevant command is executed within the 'backrest-repo'
pod. As such, this PR separates out the check for the Primary pod
from the 'show backup' functionality so that the backup information
can be shown regardless of the Primary pod's status.