Update standard image tag when using a GIS enabled cluster
Currently, the image tag value used by GIS enabled pgclusters is
being used when creating sidecontainers. Since the standard and
GIS enable PostgreSQL containers use different tags, this causes
an image pull error.

This update takes the current image name and the image tag value
stored in the pgcluster CRD and, if the image being used is the
crunchy-postgres-gis-ha container with the corresponding tag, it uses
an updated tag without the addition of the GIS version when provisioning
sidecars containers for the cluster.

Issue: [ch9393]
Issue: #1749