Create "pgo-config" ConfigMap if not provided
The Operator (and associated containers) will create the pgo-config
ConfigMap from the stored default configuration if the pgo-config
ConfigMap is not present.

This allows for the editing of "default configuration" after
the Operator is deployed, and allows for the removal of the
pgo-config creation step from several of the installation
methods.

Issue: [ch9451]