Preserve cluster-level resources in a multi-Operator environment
Currently, uninstalling an instance of the PostgreSQL Operator
from one namespace will remove all Kubernetes/Openshift cluster-
level resources as well. This impairs the functionality of any other
Operators that may be installed.

This update adds a check to see if there are any other PostgreSQL
Operators installed on the cluster and, if so, adds a flag to skip
the removal of cluster-level resources.