Use autogenerated pgouser password if not set
Currently, the PostgreSQL Operator requires a pgouser's
password to be set during the installation process.

This update allows for the password to be unset during installation,
and, during Operator initialization, any pgouser that currently has an
empty password will be given an autogenerated password, stored in the
relevant pgouser secret.