Trace Kubernetes API calls using OpenTelemetry
OpenTelemetry offers instrumentation packages that are agnostic to the
tool(s) that finally store telemetry data. When the SDK for Go reaches
GA, we can leverage the OTLP exporter to offload further configuration
to an external OpenTelemetry Collector.

For now, traces can be sent to one of a handful of destinations chosen
by environment variable: Jaeger agent, Jaeger collector, stdout as JSON,
or a file as JSON.

Issue: [ch9735]