Initial Controller for postgrescluster CR
Adds the initial controller for the 'postgrescluster'  custom resource.
This includes the creation of a controller runtime 'manager', which is
responsible running and managing the 'postgrescluster' controller (and
will ultimately manage any additional controllers created in support of
managing PostgreSQL clusters via the 'postgrescluster' custom
resource).

The controller itself is created using the controller runtime 'builder'
package, and includes a reconcile function that will be responsible for
reconciling PostgreSQL clusters via events generated for
'postgrescluster' custom resources.  For this initial controller
implementation, the reconcile function simply logs a debug message when
a 'postgrescluster' event is received, providing a structure and layout
for adding the business logic to reconcile an actual PostgreSQL cluster
in subsequent commits.  Also, the controller demonstrates how the
controller runtime client can be utilized to obtain a 'postgrescluster'
resource from the Kubernetes API, while also showing how native
Kubernetes events can be generated for 'postgrescluster' custom
resources.

Also included are kustomize artifacts and associated specifications
for Deploying the PostgreSQL Operator to run the new  controller
runtime 'manager' discussed above, and therefore the new
'postgrescluster' controller.  This includes the various, namespaces,
RBAC, etc. needed to support the Deployment of the operator, along
with the Deployment artifact for the PostgreSQL Operator itself.  Make
targets have also been created to facilitate deploying the PostgreSQL
Operator via kustomize (e.g. 'make deploy', 'make install', etc.).
Additionally, a simple example has been created to create a
'postgrescluster' custom resource using kustomize.

It should be noted that with this initial implementation the
'postgres-operator' ServiceAccount is assigned a ClusterRole to
to provide it with the ability manage 'postgrescluster' resources in
any namespace.  However, by using the 'PGO_TARGET_NAMESPACE'
environment variable, the PostgreSQL Operator can be configured to
only listen for 'postgrescluster' resources in a specific namespace.

And finally, environment variables have been created to allow either
the new 'postgrescluster' controller, or the current 'pgcluster'
controller (along with supporting controllers, e.g. for pgtask,
pgreplica, etc.) to be disabled.  This allows the PostgreSQL Operator
to run the controllers for either PostgreSQL cluster custom resource
as needed (including the ability to run both at the same time if
desired).  The kustomize installers for the new 'postgrescluster'
custom resource are designed to only enable the  'postgrescluster'
controller, while the installers for the 'pgcluster' custom resource
have been updated to only enable to 'pgcluster' controllers.

Issue: [ch9475]