Compaction of pgo-sqlrunner into crunchy-postgres
As part of the compaction effort the pgo-sqlrunner is now a running mode in the
crunchy-postgres image. The pgo-sqlrunner image has been remove and related
files have been moved to the Crunchy Containers repo. References to the image
have been updated to use the crunchy-postgres image and the running mode `MODE:
sqlrunner`.