Updates to use compacted crunchy-postgres image
This change updates code and templates that reference any old image that is now
part of the compacted crunchy-postgres image. Templates have been updated to
use crunchy-postgres as the image and pass in the relevant running mode env
variable. The code has been updated to use the new `/opt/crunchy` path instead
of `/opt/cpm`.

The crunchy-pgdump and crunchy-pgrestore images were compacted into the
crunchy-postgres image. This change removes references to the old images from
pull scripts and images.go. The update to images.go removes the related images
for each of the compacted images and updates the code to use the related image
for crunchy-postgres.