Add TLS support for pgBouncer
Since 2b05f01b, the Postgres Operator has supported configuration of TLS
connections for PostgreSQL clusters. However, connections made to
pgBouncer did not support TLS directly, though one could do some heavy
work to the template files to bring about this support.

This introduces the ability to directly configure pgBouncer instances
with TLS, with the following preconditions:

- TLS MUST be enabled within the PostgreSQL cluster
- pgBouncer and the PostgreSQL share the same certificate authority (CA)
bundle

When TLS is enabled, connections are facilitated with the following
default rules:

- All connections to pgBouncer MUST be over TLS. Effectively, this
is "TLS only" if connecting via pgBouncer.
- Connections coming into pgBouncer have a PGSSLMODE of "require"
- Connections going into PostgreSQL a PGSSLMODE of "verify-ca"

This adds an attribute to the pgcluster Spec in the "pgBouncer" section
called "tlsSecret", which will store the name of the TLS secret to use
for pgBouncer.

One can also enable TLS for pgBouncer in the following ways:

- `pgo create cluster --pgbouncer-tls-secret` if `--pgbouncer` is set
- `pgo create pgbouncer --tls-secret`

Issue: [ch9777]