Catch error when generating default pgo-config
During the initialization of the default "pgo-config" ConfigMap,
there exists a case (likely a race condition that I did not track
down) that triggered an error, but we were not catching the error.

Regardless, we should, given the next line could trigger a panic.
Immediate remediation without the patch should be just restarting
the Operator Pod.

Issue: [ch9826]
Issue: #2075