Introduce rolling update interface for PostgreSQL clusters
A rolling update of a PostgreSQL cluster involves applying any updates
that may require downtime to each replica within a PostgreSQL cluster,
followed by the promotion to a replica deemed suitable to be a primary,
followed by the update being applied to the former primary.

This commit introduces an interface to perform this exact behavior, by
allowing for any updates to the Deployments of PostgreSQL instances to
have any updates applied in a rolling fashion.

Issue: [ch9881]