Modify cluster update annotations to use rolling updates
Modifying the annotations on the template portion of a Deployment
Spec causes each Pod that is under management of a Deployment to
be restarted. For managing a database server, this can be less than
ideal.

As such, it is prudent to employ a rolling update strategy for
annotation updates on database instances to minimize downtime on
the primary instance.