Remove disabling of autofailover from rmdata
The rmdata application no longer disables autofailover when deleting a
PostgreSQL cluster.  In past versions of the PostgreSQL Operator, it
was necessary to first disable autofailover prior to cluster deletion
since the rmdata application itself would stop the PostgreSQL database
using 'pg_ctl stop'.  However, since Patroni is now responsible for
cleanly shutting down the database (specifically upon receipt of a
SIGTERM signal), autofailver should no longer be disabled (if it is,
Patroni will not respond to the SIGTERM and will therefore not attempt
to cleanly shutdown the database).

This commit therefore ensures an attempt is made to cleanly shutdown
the database when deleting a PostgreSQL cluster.  This, in turn, will
increase the likeliness that the cluster can later be recreated and
cleanly restarted.

Issue: [ch9856]