Delta Restore for In-Place Cluster Restore
When performing an in-place PostgreSQL cluster restore (such as when
using the 'pgo restore' command), the PVC for the current primary
database (includes the PGDATA PVC, along with any WAL and/or tablespace
PVCs) will now be preserved if found (as identified by the 'current-
primary' annotation on the pgcluster custom resource).  This will cause
the 'crunchy-postgres-ha' container to attempt a pgBackRest "delta"
restore when bootstrapping the restored cluster, therefore leveraging
any existing data within the PGDATA directory efficiently restore the
database.

Issue: [ch9878]