Add `--rolling` flag to `pgo restart`
The `--rolling` flag allows for one to specify a restart on a
PosgreSQL cluster to occur in a rolling fashion, i.e. all the
replicas a restarted, then a switchover occurs, then the newly
demoted primary is restarted.

This subsequently creates a task custom resource to perform the
rolling update, as said updates can take some time to process.

Issue: [ch9881]