pgo-backrest and pgo-backrest-repo containers cleanup
The pgo-backrest and pgo-backrest-repo containers have been
moved to the Crunchy Containers project. As such, the associated
files are no longer needed in this repository.

Additionally, the references to these containers are now updated
to match the new naming convention being used, and the image
tag and prefix values are updated to reflect the new location
of the containers.

This change removes debug flags and references to the unused
sshd_port env variable. It also fixes a minor copy paste error in
the docs.

Co-authored-by: TJ Moore <tj.moore@crunchydata.com>