Revert label change
As part of the compaction changes some labels and label checks were changed.
This pr reverts these changes.

The `pgo-scheduler` code was updated to check for a `crunchy-pgbackrest-repo` label
instead of the `pgo-backrest-repo` label. The deployment templates were not
updated to use the updated label so the scheduler would fail to create a backup
job when scheduling a backup.

the pgo.sqlrunner template was updated to have the `sqlrunner` label instead of
`pgo-sqlrunner`