Aggregate rolling update triggered behavior
Updates to a PostgreSQL cluster that warrant a rolling update
are now aggregated to only trigger a single rolling update per
action taken on a PostgreSQL cluster. This allows for the changes
to be rolled out more rapidly, as well as limit the number of
downtime events that need to take place.