Improve compatibility between templates and Kubernetes objects
The updates the "exporter.json" template, which is used for deploying
the "crunchy-postgres-exporter" sidecar for metrics collection in a
PostgreSQL cluster, to not have a preceiding "," in it. This in turn
allows for the file to be mapped into a Kubernetes Container object,
for convenience of manipulation within a program.