Enable / disable the metrics sidecar during cluster lifetime
This commit introduces the ability to enable/disable the metrics
collection sidecar (`crunchy-postgres-exporter`) during the lifetime of
a PostgreSQL cluster. This can be toggled in multiple ways, including:

- The `exporter` attribute in pgclusters.crunchydata.com
- `pgo update cluster --enable-metrics`, which adds the sidecar
- `pgo update cluster --disable-metrics`, which removes the sidecar

As adding/removing a sidecar results in modifying a Deployment template,
this action will trigger a rolling update of the PostgreSQL cluster in
an effort to minimize any downtime.

This also has the net effect of moving the "ccp_monitoring" used that is
created to being fully managed by the Postgres Operator. The
`CollectSecretName` attribute is now removed from the pgcluster CRD, as
is the "PgMonitorPassword" attribute from the `pgo-deployer` and other
installers.

Issue: [ch7270]
Issue: #1413