Wait for PG Deployment Deletion on Restore
Since the primary PVC for the cluster is now retained during an
in-place PostgreSQL cluster restore in support of pgBackRest delta
restores, when preparing a cluster for a restore we can no longer rely
on the deletion of all PVC's as an indicator that the 'config' and
'leader' ConfigMaps created by Patroni can be removed.  Therefore,
the Operator now specifically waits for all Deployments to be
successfully removed prior to deleting these resources.

Issue: [ch9926]