Ensure consistent permissions for mounting repo Secret
While the Secret volume mount is set to be readonly for the
pgBackRest Secret information, the defaultMode on the volume itself
was set to be more permissive. While it appears that the vast
majority of Kubernetes distributions gie precedence to the value
of the volume mount, so flavors do use the values set on the volume.

As such, it's prudent to remove the more permissive settings, which
this patch does.

Issue: [ch9989]
Issue: #2140