Allow for explicit deletion of pgBackRest backups
This introduces the ability to delete pgBackRest backups using
the `pgo delete backup` command. The pgBackRest backup name
must be specified using the `--target` command, which can be
determined through a call to `pgo show backup`.

This also includes obligatory language on when to use explicit
backup deleting to ensure the user does not take actions on
their pgBackRest repository that they do not intend to.

Issue: #2111