Update linter CI to include "default" as exhaustive
There are some switch statements that may have a larger amount
of values that can be considered, and as such we should allow
for default to mean all of them.

This now also runs on the full code base each time a pull request
is opened. Additionally, it disables the gofumpt check.