Update Init Flag if PGHA ConfigMap Already Exists
If an existing PGHA ConfigMap is identified when bootstrapping a new
PostgreSQL cluster, the Operator now updates the "init" flag within
the existing ConfigMap (specifically by setting it to "true").  This
ensures proper configuration of the "init" flag (and therefore the
proper functionality of any logic within the Operator that relies on
it, e.g. the execution of init logic within the 'crunchy-postgres-ha'
container) when the PGHA ConfigMap is pre-defined prior to PostgreSQL
cluster initialization.

Issue: [ch9986]