Rewrite `pgo scale` endpoint to accept POST requests
This rewrites the `pgo scale` API endpoint (/custers/scale/{name})
to accept POST instead of GET requests, and moves all of the
parameters to the POST body. However, the "{name}" parameter in
the URL path still takes precedence in reference to the cluster
name.

This change allows for more complex parameters to be passed into
the scale request, as well as moves the request itself to something
that is closer to the REST convention.