Allow for PostgreSQL user Secrets to be generated by Operator
Previously, the PostgreSQL user Secrets required for bootstrap
(superuser, replication user, standard user) needed to be present
before adding a custom resource. Now, the PostgreSQL Operator can
create the missing Secrets before bootstrapping a cluster.

The API server behavior does not change, as the API server can
both create the user credentials and return them to the end user.

This also modifies the examples to not include a requirement to
generate the replication user credential. As this is really a
managed service credential, there is no reason to put this onus
on the user. The superuser credential is still included in the
exmaples, as there are certain situations where one should log
in as the superuser.