Eliminate user Secret name attributes from pgcluster CRD
These Secrets have been fully managed by the Postgres Operator
for a long time and there is no reason why they need to be specified
by the user as specific custom resource attributes. This further
simplifies the number of steps required to create a PostgreSQL
cluster with the operator through a "custom resource" workflow.