Allow for pgBadger sidecar to be enabled/disabled
Similar to the metrics collection sidecar, the pgBadger sidecar
can now be enabled during the lifetime of a PostgreSQL cluster.
This also adds the `--enable-pgbadger` and `--disable-pgbadger`
flags to the `pgo update cluster` command.

Additionally, the update to the pgBadger sidecar can be triggered
through a modification to the `pgclusters.crunchydata.com` CR,
and will trigger a rolling update to minimize downtime.