Move autofail toggle to `DisableAutofail` attribute on CRD
This required relatively little shifting of code, as this former
label effectively behaved like an attribute. Given autofail (read:
HA) is the default in the PostgreSQL Operator, the chosen nomenclature
is to reflect that we want the use to make a conscious decision to
disable HA (given Go defaults booleans to "false").