Move pgBackRest storage type to CRD attribute
This previously was driven by a label stored on a
pgclusters.crunchydata.com custom resource. However, this is structured
data, as well as data that drives cluster behavior, and as such it
should be moved. This adds a new attribute called "backrestStorageTypes"
which contains a list of acceptable storage types, i.e.:

- posix
- s3

"local" is kept for backwards compatibility purposes, but the upgrade
attempts to force things to "posix". Additionally, if left empty,
"posix" is assumed by the Postgres Operator.

Modifying "backrestStorageTypes" on an existing cluster currently has
no effect, though this is set up to be able to switch pgBackRest
repositories during the lifetime of a cluster.

These changes also have the effect of simplifying "pgo backup" behavior.
If the command "pgo backup <clusterName>" is called, it will
automagically issue backups to all available pgBackRest repositories.
For example, a cluster that uses "s3" or "posix"/"s3" repositories will
automatically have backups issued to them without specifying a target.
The ability to specify taking a backup to a specific target still
remains.

The same applies to the restore in place functionality, as the default
storage type is known.