Ensure pod anti-affinity is loaded from spec
The pod anti-affinity had been driven from the pgcluster custom
resource since its inception, but there were some vestiges
in the form of labels on the custom resource. This moves all of
the labelign to be driven by the custom resource attributes, which
primarily affects the Postgres instances.

This also includes some random documentation updates.