Move node affinity controls to CRD attributes
Node affinity had been controlled by the use of an amorphous label
on the `pgclusters.crunchydata.com` and `pgreplicas.crunchydata.com`
attributes. This is now moved to use a similar approach on the CRD
as used for pod anti-affinity, though using the standard Kubernetes
definition for setting node affinity on a Pod.

The `--node-label` syntax on the various pgo-client commands still work
as expected and are mapped to using the new format. The attributes
on the CRDs will allow for expansion over the number of supported node
affinity rules that can be placed on a PostgreSQL cluster, particularly
as Kubernetes deployment topologies continue to evolve.