Allow for selection of node affinity type from pgo client
The previous commit allows for node affinity type (e.g. preferred, required) to
be selected directly from a custom resource. This is now possible from the
command line using the `--node-affinity-type` flag, which accepts values of
"preferred" and "required" ("prefer" and "require" are also accepted).

This flag is available on the following commands:

- `pgo create cluster`
- `pgo scale`
- `pgo restore`

Note that this flag needs to be used in conjunction with the `--node-label`
flag.