Add `--service-type` flag to `pgo update cluster`
While a previous commit introduced the ability to update the Service
type of a PostgreSQL cluster via a custom resource, this commit
introduces the ability to do so from the pgo client.

Additionally, this ensures a Service Type update to a cluster propagates
to pgBouncer if the pgBouncer Service type is set to empty.