Add support for toleration updates using pgo client
This commit adds the `--toleration` flag to the `pgo update cluster`
command and supports to adding and removing of tolerations to an
existing PostgreSQL cluster.

A toleration can be added in similar fashion to the --toleration
flag when running `pgo create cluster` or `pgo scale`.

A toleration can be removed by appending a `-` to the end of the
toleration flag value, e.g.

    pgo update cluster hippo --toleration=zone=east:NoSchedule-