Enable user-defined forced failovers
While this was the default behavior in Operator's past when using
`pgo failover`, a previous commit changed the Operator internals
to leverage a "controlled switchover" which is a bit nicer (i.e.
it only works if there is a healthy instance to fail over to).
However, there are situations where one must force a failover,
and as such, we need to allow for one to do so.

This adds the `--force` flag to `pgo failover` to allow for a
forced failover. Note that `--target` must explicitly be set
when forcing a failover.