Initial reconcile for Patroni instances
The PostgresCluster controller creates the ConfigMaps, Services, and
StatefulSets necessary to run Patroni for a PostgresCluster object.
These are deleted when the PostgresCluster is deleted.

The StatefulSets are configured so that each of their Pods gets a stable
DNS name, and Patroni is configured to connect to other Patroni members
using those names. This should make it easier to manage TLS certificates
for Patroni.

Issue: [ch9791]