Deleting a pgcluster resource deletes a PostgreSQL cluster
This behavior, which had been removed circa 4.1, is now reintroduced
to use the present machienry when a PostgreSQL cluster is being
deleted. Deleting a pgcluster resoruce directly will now delete a
PostgreSQL cluster and all of its associated objects.

To retain the PVCs for the PostgreSQL data directory or the backups
repository, one can set the following Annotations on a pgclusters
custom resource:

- `keep-backups`: indicates to keep the pgBackRest PVC when deleting
the cluster.
- `keep-data`: indicates to keep the PostgreSQL data PVC when deleting
the cluster.

Issue: [ch9435]
Issue: #1203