Tighten up ephemeral storage limits
This removes an ephemeral storage volume that is no longer used
and places a cap of 64Mi on the pgBadger ephemeral storage
is not mounted unless there is a pgBadger instance directory.

This does not add a size limit to shared memory (dshm), as that
is a much more complicated topic.

Issue: #2188