Bump sigs.k8s.io/controller-runtime to v0.7.0
The Kubernetes version installed by envtest is now v1.19.2, which
includes fixes to the handling of managed fields in subresources.

This released version uses logr.FromContext and logr.NewContext, so we
need to be sure to pass our singleton to its constructors. We can now
see log fields that it sets as well.