Update perms on deployer working directory
The working directory for pgo-deployer (/tmp/.pgo) is now created at build time
instead of install time. When building, the directory is created with root:root
as user:group meaning the installer does not have permission to create the
required subdirectories. This change recursively updates the user and group on
/tmp/.pgo so that the installer can make changes.