Change Patroni scope to avoid naming conflicts
When using Kubernetes Endpoints for Patroni's DCS, the leader lease has
the same name as the scope. We had used the cluster name for the scope,
to leverage the same labels we put on all our objects. Unfortunately,
this combination means it's possible to create two clusters that will
try to write to the same DCS objects.

This commit adds a suffix to the Patroni scope to avoid any naming
collisions with our own objects. A new test makes it easy to ensure this
holds true going forward.

This also replaces the Service that we expose to users for connecting
to the primary PostgreSQL instance. This new Service can be named and
configured independently of the Patroni scope and Patroni DCS. The
PostgreSQL port is also now plumbed correctly.

Issue: [ch10151]