Create PKI interface for managing Operator managed certificates
This establishes the foundational layer for a public key infrastructure
(PKI) for the Postgres Operator to manage identity and verification
across all managed PostgreSQL clusters and associated instances. For
this aspect, this includes creating a three layer system of certificate
authorities (CAs) and certificates for the different management layers
of the PostgreSQL Operator.

This includes:

- A root certificate authority (root CA). This is the primary CA for
the PostgreSQL Operator, and should be stored in the same namespace that
the PostgreSQL Operator is deployed to.

- An intermediate certificate authority (intermediate CA). An
intermediate CA should be deployed to each namespace that is watched by
the PostgreSQL Operator. Note that even in a single namespace
environment, there should still an intermediate CA to help facilitate
certificate rotation.

- A leaf certificate, which can be distributed as the identity for the
clusters and other accessible services (e.g. pgBouncer) that the
PostgreSQL Operator manages.

This includes support for PEM encryption of private keys.

Private keys are generated using the ECDSA algorithm with a P-256 curve
with a SHA-384 digest. The P-256 curve is recommended by Mozilla for
supporting modern TLS systems:

  https://wiki.mozilla.org/Security/Server_Side_TLS

and the SHA-384 digest provides resistance against length extensions
attacks.

For PEM encryption we use AES-256 encryption with the CBC cipher mode,
which is provided by the standard Go X509 library.

Issue: [ch9806]