Ensure optimistic locking on controller references
It is recommended that a controller not take any action on an object it
does not own. When using server-side apply, we rarely have a copy of an
object before we write so we (1) do not check for existing ownership and
(2) send no resourceVersion for optimistic concurrency control.

We want to be certain that server-side apply works as documented when it
comes to controller references.

Tested on Kubernetes v1.16.4, v1.17.9, and v1.19.2.

Issue: [ch10223]
See: https://git.k8s.io/community/contributors/design-proposals/api-machinery/controller-ref.md