Add support for the pgBackRest `--compress-type` flag
Selecting the compression type for pgBackRest is supported across
recent Postgres Operator releases, but the CLI was not allowing for
this flag to be passed through.

This allows for the selection of pgBackRest compression type amongst
the allowable methods in the container, which are none, bz2, gz, and
lz4. Currently zst does not ship within the container.

Issue: [ch10287]