Default SSH configuration
Adds the default SSH configuration as required for
connection to and from the pgBackRest repo pod.

To facilitate this, a configmap that holds the required
configuration files and secret containing the ecdsa key
pair. For a postgrescluster named 'mycluster', the configmap
will be named 'mycluster-ssh-config' and the secret will be
named 'mycluster-ssh'. These are then used for the
creation of a projected volume and associated volume mount
to be used by the various pods.

In the configmap, the SSH client and SSHD service configuration
files will be found at the 'config' and 'sshd_config' key
references.

In the secret, the private key will be found at the 'id_ecdsa'
key reference. The public key will be found at the 'id_ecdsa.pub'
key reference.

Within the containers themselves, the file references will be located
under the '/sshd' path. As such, the command to start the sshd daemon
will be similar to

/usr/sbin/sshd -eD -f sshd/sshd_config

Similarly, the ssh client command will be similar to

ssh <ip> -F /sshd/config

Otherwise, no other scripting should be required.

Finally, the various pgBackRest configuration go tests have been
refactored so that duplicated functions and constants are now located
in the test_helpers.go file and an unneeded runtime manager has been
removed.